Update Feature
User can left empty the unknown value, not only with 0
